requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
google-genai==1.0.0
tqdm==4.66.4
